requirements.txt
PyGithub==2.6.1
PyYAML>=6.0.1
requests>=2.32.2
python-dotenv>=1.0.0
